org-manager/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
pyjwt==2.8.0
bcrypt==4.0.1
python-multipart==0.0.6
pandas==2.1.3
openpyxl==3.1.2
pydantic==2.5.0
pydantic-settings==2.1.0
